[android] #2579 - Moving InvalidAccessTokenException to the exceptions package
